requirements.txt
requests>=2.32.0
pydub
PyYAML>=6.0.1
feedparser>=6.0.11
python-dateutil>=2.9.0
python-dotenv>=1.0.0
tenacity>=8.2.3
python-slugify>=8.0.4
boto3>=1.34.0
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
crawl4ai>=0.8.0
playwright>=1.50.0
pytest>=8.0.0
